refactor(web): dedup VLAN and port range parsing into shared util
Move parseRangeStr/parseRangesRaw into utils/ranges.ts and have the forward rule drawer reuse it instead of its own copy.
